pyproject.toml
[project]
name = "cse-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "dash>=3.0.4",
    "dash-bootstrap-components>=2.0.2",
    "fastapi>=0.115.12",
    "fastparquet>=2024.11.0",
    "google-genai>=1.13.0",
    "mistralai>=1.7.0",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "structlog>=25.3.0",
    "tenacity>=9.1.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
